test: cover docs prerelease dispatch
