+++ cyborg/package.json
@@ -4,6 +4,9 @@
   "description": "Automator for GreptimeDB Repository Management",
   "private": true,
   "packageManager": "pnpm@8.15.5",
+  "scripts": {
+    "test": "tsx --test src/*.test.ts"
+  },
   "dependencies": {
     "@actions/core": "^1.10.1",
     "@actions/github": "^6.0.0",
